Refactor frameworks package
