+++ src/main/java/uk/gov/hmcts/reform/prl/exception/ControllerExceptionHandler.java
@@ -9,7 +9,6 @@ import org.springframework.web.ErrorResponse;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-import uk.gov.hmcts.reform.prl.framework.exceptions.DocumentGenerationException;
 
 @RestControllerAdvice
 @Slf4j
