+++ src/main/java/uk/gov/hmcts/reform/prl/exception/WorkflowException.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.prl.framework.exceptions;
+package uk.gov.hmcts.reform.prl.exception;
 
 public class WorkflowException extends Exception {
