+++ src/main/java/uk/gov/hmcts/reform/prl/models/caseflags/AllPartyFlags.java
@@ -6,7 +6,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.extern.jackson.Jacksonized;
-import uk.gov.hmcts.reform.prl.models.caseflags.Flags;
 
 @Data
 @Builder(toBuilder = true)
