+++ src/main/java/uk/gov/hmcts/reform/prl/models/documents/OtherDocument.java
@@ -3,7 +3,6 @@ package uk.gov.hmcts.reform.prl.models.documents;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import lombok.Builder;
 import lombok.Data;
-import uk.gov.hmcts.reform.prl.models.documents.Document;
 
 @Data
 @Builder
