+++ src/main/java/uk/gov/hmcts/reform/prl/models/tasklist/Task.java
@@ -3,7 +3,6 @@ package uk.gov.hmcts.reform.prl.models.tasklist;
 import lombok.Builder;
 import lombok.Data;
 import uk.gov.hmcts.reform.prl.enums.Event;
-//import uk.gov.hmcts.reform.prl.enums.TaskState;
 
 import java.util.Optional;
