+++ src/main/java/uk/gov/hmcts/reform/prl/tasks/emails/ExampleEmailTask.java
@@ -2,12 +2,12 @@ package uk.gov.hmcts.reform.prl.tasks.emails;
 
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.prl.enums.LanguagePreference;
-import uk.gov.hmcts.reform.prl.framework.context.TaskContext;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.CaseDetails;
 import uk.gov.hmcts.reform.prl.models.dto.notify.CitizenEmail;
 import uk.gov.hmcts.reform.prl.models.dto.notify.EmailTemplateVars;
 import uk.gov.hmcts.reform.prl.models.email.EmailTemplateNames;
 import uk.gov.hmcts.reform.prl.services.EmailService;
+import uk.gov.hmcts.reform.prl.tasks.TaskContext;
 import uk.gov.hmcts.reform.prl.tasks.emails.generics.SendEmailTask;
 
 @Component
