+++ src/main/java/uk/gov/hmcts/reform/prl/workflows/ExampleWorkflow.java
@@ -3,10 +3,9 @@ package uk.gov.hmcts.reform.prl.workflows;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
-import uk.gov.hmcts.reform.prl.framework.exceptions.WorkflowException;
-import uk.gov.hmcts.reform.prl.framework.task.Task;
-import uk.gov.hmcts.reform.prl.framework.workflow.DefaultWorkflow;
+import uk.gov.hmcts.reform.prl.exception.WorkflowException;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.CaseDetails;
+import uk.gov.hmcts.reform.prl.tasks.Task;
 import uk.gov.hmcts.reform.prl.tasks.emails.ExampleEmailTask;
 
 @Component
