+++ src/main/java/uk/gov/hmcts/reform/prl/workflows/ValidateMiamApplicationOrExemptionWorkflow.java
@@ -3,10 +3,9 @@ package uk.gov.hmcts.reform.prl.workflows;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
-import uk.gov.hmcts.reform.prl.framework.exceptions.WorkflowException;
-import uk.gov.hmcts.reform.prl.framework.task.Task;
-import uk.gov.hmcts.reform.prl.framework.workflow.DefaultWorkflow;
+import uk.gov.hmcts.reform.prl.exception.WorkflowException;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.WorkflowResult;
+import uk.gov.hmcts.reform.prl.tasks.Task;
 import uk.gov.hmcts.reform.prl.tasks.ValidateMiamApplicationOrExemptionTask;
 
 @Component
