+++ src/test/java/uk/gov/hmcts/reform/prl/services/validators/RespondentRelationshipCheckerTest.java
@@ -12,7 +12,7 @@ import uk.gov.hmcts.reform.prl.models.complextypes.RespondentRelationOptionsInfo
 import uk.gov.hmcts.reform.prl.models.dto.ccd.CaseData;
 import uk.gov.hmcts.reform.prl.services.TaskErrorService;
 
-import  static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertTrue;
 import static org.testng.AssertJUnit.assertNotNull;
 
 @RunWith(MockitoJUnitRunner.class)
