+++ src/test/java/uk/gov/hmcts/reform/prl/utils/TaskContextProvider.java
@@ -1,8 +1,8 @@
 package uk.gov.hmcts.reform.prl.utils;
 
 import lombok.NoArgsConstructor;
-import uk.gov.hmcts.reform.prl.framework.context.DefaultTaskContext;
-import uk.gov.hmcts.reform.prl.framework.context.TaskContext;
+import uk.gov.hmcts.reform.prl.tasks.DefaultTaskContext;
+import uk.gov.hmcts.reform.prl.tasks.TaskContext;
 
 @NoArgsConstructor
 public class TaskContextProvider {
